EVERYONE PLZ TAKE A LOOK
Clicking buttons in the the tests don't work because of the animations (I think). I added .animationEnd() to button click, it worked beautifully once, and then now it doesn't recognize it as a function. I don't know what's going on and I feel like my brain is gonna explode, so yall plz take a look. I probably forgot a semicolon or something dumb. Thank you and bless up.